# Notes-Server/swagger/swagger.json
{
  "openapi" : "3.0.0",
  "servers" : [ {
    "url" : "http://localhost:3000/api/v1"
  } ],
  "info" : {
    "description" : "Add, edit, manager and share notes between users",
    "version" : "0.1-oas3",
    "title" : "Notes"
  },
...
  "paths" : {
...
    "/api/v1/groups/create" : {
      "post" : {
        "tags" : [ "groups" ],
        "summary" : "Create a new group",
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "403" : {
            "description" : "Not authenticated",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          }
        },
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "properties" : {
                  "userid" : {
                    "type" : "string"
                  },
                  "title" : {
                    "type" : "string"
                  },
                  "description" : {
                    "type" : "string"
                  }
                },
                "type" : "object"
              }
            }
          },
          "required" : true
        }
      }
    },
    "/api/v1/groups/delete" : {
      "post" : {
        "tags" : [ "groups" ],
        "summary" : "Delete a group",
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "403" : {
            "description" : "Not authenticated",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "404" : {
            "description" : "Cannot find group",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          }
        },
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "properties" : {
                  "userid" : {
                    "type" : "string"
                  },
                  "groupid" : {
                    "type" : "string"
                  }
                },
                "type" : "object"
              }
            }
          },
          "required" : true
        }
      }
    },
    "/api/v1/groups/addnote" : {
      "post" : {
        "tags" : [ "groups" ],
        "summary" : "Add a note to a group",
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "403" : {
            "description" : "Not authenticated",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "404" : {
            "description" : "Cannot find note/Cannot find group",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          }
        },
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "properties" : {
                  "userid" : {
                    "type" : "string"
                  },
                  "groupid" : {
                    "type" : "string"
                  },
                  "noteid" : {
                    "type" : "string"
                  }
                },
                "type" : "object"
              }
            }
          },
          "required" : true
        }
      }
    },
    "/api/v1/groups/removenote" : {
      "post" : {
        "tags" : [ "groups" ],
        "summary" : "Remove a note from a group",
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "403" : {
            "description" : "Not authenticated",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "404" : {
            "description" : "Cannot find note/Cannot find group",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          }
        },
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "properties" : {
                  "userid" : {
                    "type" : "string"
                  },
                  "groupid" : {
                    "type" : "string"
                  },
                  "noteid" : {
                    "type" : "string"
                  }
                },
                "type" : "object"
              }
            }
          },
          "required" : true
        }
      }
    },
    "/api/v1/groups/" : {
      "post" : {
        "tags" : [ "groups" ],
        "summary" : "Get all groups created by a user",
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "responses" : {
          "200" : {
            "description" : "success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Group"
                  }
                }
              }
            }
          },
          "403" : {
            "description" : "Not authenticated",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "404" : {
            "description" : "No groups found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          }
        },
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "properties" : {
                  "userid" : {
                    "type" : "string"
                  }
                },
                "type" : "object"
              }
            }
          },
          "required" : true
        }
      }
    },
    "/api/v1/groups/group" : {
      "post" : {
        "tags" : [ "groups" ],
        "summary" : "Get details of a particular group",
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "responses" : {
          "200" : {
            "description" : "success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Group"
                }
              }
            }
          },
          "403" : {
            "description" : "Not authenticated",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "404" : {
            "description" : "No group found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          }
        },
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "properties" : {
                  "userid" : {
                    "type" : "string"
                  },
                  "groupid" : {
                    "type" : "string"
                  }
                },
                "type" : "object"
              }
            }
          },
          "required" : true
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "Response" : {
        "properties" : {
          "message" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "Note" : {
        "properties" : {
          "noteId" : {
            "type" : "string"
          },
          "userId" : {
            "type" : "string"
          },
          "permission" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "content" : {
            "type" : "string"
          },
          "sharedNotes" : {
            "type" : "array",
            "items" : {
              "type" : "object",
              "properties" : {
                "noteId" : {
                  "type" : "string"
                },
                "userId" : {
                  "type" : "string"
                },
                "permission" : {
                  "type" : "string"
                }
              }
            }
          },
          "parentNoteId" : {
            "type" : "string"
          },
          "groupId" : {
            "type" : "string"
          },
          "remainderId" : {
            "type" : "string"
          },
          "favorite" : {
            "type" : "boolean"
          },
          "createdOn" : {
            "type" : "string"
          },
          "modifiedOn" : {
            "type" : "string"
          }
        }
      },
      "Group" : {
        "properties" : {
          "groupId" : {
            "type" : "string"
          },
          "userId" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "notes" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Note"
            }
          },
          "createdOn" : {
            "type" : "string"
          },
          "modifiedOn" : {
            "type" : "string"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}